Solution to Calculate Average Marks of Students

+++ Calculate Average Marks of Students@@67877d61dc107486ad5c2743/code.c
@@ -18,7 +18,7 @@ int main() {
         scanf("%d %s %f", &students[i].roll_number, students[i].name, &students[i].marks);
         sum+=students[i].marks;
     }
-    printf("Average Marks: %.2f",sum/(float)N);
+    printf("Average Marks: %.2f",sum/N);
 
     return 0;
 }
